스케줄 수정/삭제 시 소유권 체크 누락 (#26)
* 새로운 예외 클래스 추가

스케줄 처리 시 권한 부족 상황을 위한 UnauthorizedScheduleException 클래스가 추가되었습니다.

* 인증 사용자 정보를 스케줄 수정 및 삭제에 추가

스케줄 수정(update) 및 삭제(delete) 메서드에 인증 사용자 정보를 전달하도록 수정했습니다. 이를 통해 사용자별 접근 권한을 보다 명확히 처리할 수 있습니다.

* 일정 업데이트 사용자 확인

일정을 업데이트할 때 사용자 권한 검사를 추가했습니다.

* 사용자 인증 오류 테스트 추가

사용자가 자신의 스케줄이 아닌 경우 업데이트가 되지 않는지 확인하는 테스트를 추가했습니다. UnauthorizedScheduleException의 발생 여부를 검증하여 보안을 강화했습니다.

* 스케줄 삭제 시 사용자 권한 검증 추가

스케줄 삭제 시 요청한 사용자가 해당 스케줄의 소유자인지 확인하는 로직을 추가했습니다.

* 스케줄 삭제 로직에 사용자 ID 추가

스케줄 삭제 시 사용자 ID를 전달하도록 로직을 수정했습니다. 이를 통해 인증된 사용자에 따른 삭제 권한을 정확히 검증할 수 있습니다.

* 테스트: DeleteScheduleService 테스트 추가

DeleteScheduleService에 대한 테스트 코드를 추가했습니다. 사용자가 소유하지 않은 스케줄에 대해 삭제를 시도하면 UnauthorizedScheduleException이 발생하는지 확인하는 테스트를 포함했습니다.

* 불필요한 인메모리 리포지토리 초기화 제거

테스트 코드에서 불필요하게 초기화된 인메모리 리포지토리를 제거했습니다. 리포지토리는 서비스 객체를 생성하는 과정에서 이미 주입되므로 중복 초기화를 삭제하여 코드의 간결성을 개선했습니다.

* 사용자 권한 검증 로직 리팩토링

Schedule 클래스에 validateUserAuthorization 메소드를 추가하여 사용자 권한 검증 로직을 캡슐화했습니다.

* 불필요한 예외 삭제

UpdateScheduleService에서 사용되지 않는 UnauthorizedScheduleException을 제거했습니다.

* 사용자 권한 검증 로직 리팩토링

권한 검증 로직을 `validateUserAuthorization` 메서드로 이동하여 `DeleteScheduleService`의 가독성을 개선했습니다.

* setUp 메서드에서 불필요한 초기화 코드 제거

clearRepository의 인스턴스를 불필요하게 초기화하는 코드를 삭제했습니다.

* validateUserAuthorization → validateOwner

기존 메서드 이름을 더 명확히 하기 위해 변경했습니다.

+++ backend/src/main/java/github/gunkim/climbingcalendar/domain/schedule/service/DeleteScheduleService.java
@@ -1,8 +1,10 @@
 package github.gunkim.climbingcalendar.domain.schedule.service;
 
+import github.gunkim.climbingcalendar.domain.schedule.model.Schedule;
 import github.gunkim.climbingcalendar.domain.schedule.model.id.ScheduleId;
 import github.gunkim.climbingcalendar.domain.schedule.repository.ClearRepository;
 import github.gunkim.climbingcalendar.domain.schedule.repository.ScheduleRepository;
+import github.gunkim.climbingcalendar.domain.user.model.id.UserId;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,8 +13,11 @@ import org.springframework.stereotype.Service;
 public class DeleteScheduleService {
     private final ScheduleRepository scheduleRepository;
     private final ClearRepository clearRepository;
+    private final GetScheduleService getScheduleService;
 
-    public void deleteSchedule(ScheduleId scheduleId) {
+    public void deleteSchedule(ScheduleId scheduleId, UserId userId) {
+        Schedule schedule = getScheduleService.getScheduleById(scheduleId);
+        schedule.validateOwner(userId);
         scheduleRepository.deleteById(scheduleId);
         clearRepository.deleteByScheduleId(scheduleId);
     }
